chore: Rename utils file
