+++ libs/ledger-live-common/src/exchange/swap/hasParametersForSolana.ts
